+++ e2e/helpers.ts
@@ -44,7 +44,7 @@ const groupsToSkip: Set<string> = new Set(['Components/Tooltip', 'Bullet Graph']
  */
 const storiesToSkip: Map<string, string[]> = new Map(
   Object.entries({
-    'Test Cases': ['noSeries'],
+    'Test Cases': ['noSeries', 'errorBoundary'],
     Interactions: ['multiChartCursorSync'],
     'Metric (@alpha)': ['bodyContent'],
   }),