+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/xy_chart.tsx
@@ -207,7 +207,6 @@ const DEFAULT_PROPS: ReactiveChartStateProps = {
     areas: [],
     bars: [],
     lines: [],
-    points: [],
     bubbles: [],
   },
   geometriesIndex: new IndexedGeometryMap(),