+++ packages/charts/src/utils/themes/dark_theme.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import { DARK_BASE_COLORS } from './base_colors';
+import { DARK_BASE_COLORS, DARK_TEXT_COLORS, DARK_BORDER_COLORS, DARK_BACKGROUND_COLORS } from './base_colors';
 import { palettes } from './colors';
 import { Theme } from './theme';
 import { DEFAULT_CHART_MARGINS, DEFAULT_CHART_PADDING, DEFAULT_GEOMETRY_STYLES } from './theme_common';
@@ -142,7 +142,7 @@ export const DARK_THEME: Theme = {
         inner: 10,
         outer: 0,
       },
-      fill: DARK_BASE_COLORS.title,
+      fill: DARK_TEXT_COLORS.textParagraph,
       visible: true,
     },
     axisPanelTitle: {
@@ -152,12 +152,12 @@ export const DARK_THEME: Theme = {
         inner: 8,
         outer: 0,
       },
-      fill: '#D4D4D4',
+      fill: DARK_TEXT_COLORS.textParagraph,
       visible: true,
     },
     axisLine: {
       visible: true,
-      stroke: DARK_BASE_COLORS.lightShade,
+      stroke: DARK_BORDER_COLORS.borderBaseSubdued,
       strokeWidth: 1,
     },
     tickLabel: {
@@ -165,7 +165,7 @@ export const DARK_THEME: Theme = {
       fontSize: 10,
       fontFamily: DEFAULT_FONT_FAMILY,
       fontStyle: 'normal',
-      fill: '#81858f',
+      fill: DARK_TEXT_COLORS.textSubdued,
       padding: { outer: 8, inner: 10 },
       rotation: 0,
       offset: {
@@ -180,7 +180,7 @@ export const DARK_THEME: Theme = {
     },
     tickLine: {
       visible: false,
-      stroke: DARK_BASE_COLORS.lightShade,
+      stroke: DARK_BORDER_COLORS.borderBaseSubdued,
       strokeWidth: 1,
       size: 10,
       padding: 10,
@@ -188,14 +188,14 @@ export const DARK_THEME: Theme = {
     gridLine: {
       horizontal: {
         visible: true,
-        stroke: DARK_BASE_COLORS.lightShade,
+        stroke: DARK_BORDER_COLORS.borderBaseSubdued,
         strokeWidth: 1,
         opacity: 1,
         dash: [0, 0],
       },
       vertical: {
         visible: true,
-        stroke: DARK_BASE_COLORS.lightShade,
+        stroke: DARK_BORDER_COLORS.borderBaseSubdued,
         strokeWidth: 1,
         opacity: 1,
         dash: [0, 0],
@@ -219,17 +219,17 @@ export const DARK_THEME: Theme = {
   crosshair: {
     band: {
       visible: true,
-      fill: '#2a2b33', // DARK_BASE_COLORS.lightestShade,
+      fill: DARK_BACKGROUND_COLORS.backgroundBaseInteractiveHover,
     },
     line: {
       visible: true,
-      stroke: DARK_BASE_COLORS.darkShade,
+      stroke: DARK_BORDER_COLORS.borderBasePlain,
       strokeWidth: 1,
       dash: [4, 4],
     },
     crossLine: {
       visible: true,
-      stroke: DARK_BASE_COLORS.darkShade,
+      stroke: DARK_BORDER_COLORS.borderBasePlain,
       strokeWidth: 1,
       dash: [4, 4],
     },